requirements.txt
Flask==3.0.0
Werkzeug==3.0.0
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
gunicorn==21.2.0
python-dotenv==1.0.0